% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{add_draw}
\alias{add_draw}
\title{Add draw interaction to the map}
\usage{
add_draw(ol, type = "Circle", freehand = FALSE)
}
\arguments{
\item{ol}{map widget}

\item{type}{feature type; one of \code{Circle}, \code{Polygon}, \code{Point}, \code{LineString}}

\item{freehand}{logical; specifying whether or not freehand mode should be used}
}
\description{
Add draw interaction to the map
}
