% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_geojson}
\alias{add_geojson}
\title{Add vector layer to map.}
\usage{
add_geojson(ol, data = NULL, filename = NULL, style = NULL, opacity = 1)
}
\arguments{
\item{ol}{map widget}

\item{data}{geojson, ignored if \code{filename} is given}

\item{filename}{filename to read geojson from}

\item{style}{style object or \code{NULL} (use default style)}

\item{opacity}{layer opacity}
}
\description{
Add vector layer to map.
}
\examples{
\dontrun{
  geojson <- geojsonio::geojson_json(quakes[1:10, ])
  ol() \%>\% add_geojson(geojson)
}

}
