% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{add_legend}
\alias{add_legend}
\title{Add legend to the map}
\usage{
add_legend(ol, colors, labels, title = NULL, style = NULL,
  overlay = TRUE)
}
\arguments{
\item{ol}{map widget}

\item{colors}{vector of colors}

\item{labels}{vector of text labels corresponding to the colors}

\item{title}{title of the legend or \code{NULL} (no title)}

\item{style}{list of css style definition applied to the legend (container)}

\item{overlay}{display on top of the map or below?}
}
\description{
Add legend to the map
}
