% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt.R
\name{add_mapbox_vector_tiles}
\alias{add_mapbox_vector_tiles}
\title{Add mapbox vector tiles (mvt) to the map}
\usage{
add_mapbox_vector_tiles(ol, url, attribution = NULL, style = NULL)
}
\arguments{
\item{ol}{map widget}

\item{url}{url of the mvt service}

\item{attribution}{attribution}

\item{style}{style object or \code{NULL} (use default style)}
}
\description{
If you want to create and serve your own vector tiles
you may check \href{http://t-rex.tileserver.ch/}{t-rex}.
}
\examples{
\dontrun{
 # mapbox
 key <- "your_access_token"
 url <- paste0(
   "https://{a-d}.tiles.mapbox.com/v4/",
   "mapbox.mapbox-streets-v7/",
   "{z}/{x}/{y}.vector.pbf",
   "?access_token=", key
 )
 attribution <- "Add mapbox attribution here"

 ol() \%>\%
   add_mapbox_vector_tiles(url, attribution)

 # style the tiles
 ol() \%>\%
   add_mapbox_vector_tiles(
     url,
     attribution,
     style = stroke_style(color = "green") + fill_style()
   ) \%>\%
   set_view(zoom = 14)
}

}
