% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{add_select}
\alias{add_select}
\title{Add select interaction to the map}
\usage{
add_select(ol, options = select_interaction())
}
\arguments{
\item{ol}{map widget}

\item{options}{\code{\link{select_interaction}} object}
}
\description{
Globally add select interaction to the map,
which means all layers are selectable.
}
\seealso{
\code{select} parameter in \code{\link{add_geojson}}
}
