% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_tiles}
\alias{add_tiles}
\alias{add_osm_tiles}
\alias{add_stamen_tiles}
\alias{add_xyz_tiles}
\title{Add tile layers to map.}
\usage{
add_osm_tiles(ol)

add_stamen_tiles(ol, layer = "watercolor")

add_xyz_tiles(ol, xyz_url, attribution = NULL, opacity = 1)
}
\arguments{
\item{ol}{map widget}

\item{layer}{stamen layer name}

\item{xyz_url}{xyz url}

\item{attribution}{attribution (character vector)}

\item{opacity}{layer opacity}
}
\description{
Add tile layers to map.
}
\section{Functions}{
\itemize{
\item \code{add_osm_tiles}: Add osm tile layer to map

\item \code{add_stamen_tiles}: Add stamen tile layer to map

\item \code{add_xyz_tiles}: Add custom tile layer to map
}}

\examples{
\dontrun{
  xyz_url <- get_cartodb_xyz_url()
  attribution <- cartodb_attribution()
  ol() \%>\% add_xyz_tiles(xyz_url, attribution)
}

}
