% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{add_vector_data}
\alias{add_vector_data}
\title{Add vector data to map}
\usage{
add_vector_data(ol, data, ...)
}
\arguments{
\item{ol}{map widget}

\item{data}{any data object which can be parsed to geojson by
\code{\link[geojsonio]{geojson_json}}}

\item{...}{optional parameters passed to \code{\link{add_geojson}}}
}
\description{
This function is just a shorthand to \code{\link{add_geojson}} using
\pkg{geojsonio} to convert any data object to geojson.
Therefore, if the library/namespace is not available, an error is thrown.
}
