% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{add_wms}
\alias{add_wms}
\title{Add wms image layer}
\usage{
add_wms(ol, wms_url, layers)
}
\arguments{
\item{ol}{map widget}

\item{wms_url}{url of the wms service}

\item{layers}{layers}
}
\description{
Add wms image layer
}
\examples{
\dontrun {
  wms_url <- "https://ahocevar.com/geoserver/wms"

  ol() \%>\%
    add_stamen_tile() \%>\%
    add_wms(wms_url, layers = "topp:states")
}

}
