% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{add_wms_tiles}
\alias{add_wms_tiles}
\title{Add wms tile layer}
\usage{
add_wms_tiles(ol, url, layers, attribution = NULL, wms_options = NULL,
  options = layer_options())
}
\arguments{
\item{ol}{map widget}

\item{url}{url of the wms service}

\item{layers}{layers}

\item{attribution}{attribution}

\item{wms_options}{named list of wms options like \code{STYLE}, ...}

\item{options}{layer options}
}
\description{
Add wms tile layer
}
\examples{
\dontrun{
  wms_url <- "https://ahocevar.com/geoserver/wms"
  layers <- c("ne:NE1_HR_LC_SR_W_DR", "topp:states")

  ol() \%>\%
    add_wms_tiles(wms_url, layers) \%>\%
    set_view(-99.74, 32.45, zoom = 3)

  ol() \%>\%
    add_stamen_tiles() \%>\%
    add_wms_tiles("http://ows.terrestris.de/osm/service",
                  layers = "TOPO-WMS",
                  attribution = "Do not forget to insert the attribution!",
                  options = layer_options(opacity = 0.5))
}

}
