% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls}
\alias{controls}
\alias{add_scale_line}
\alias{add_mouse_position}
\alias{add_overview_map}
\title{Add controls to map}
\usage{
add_scale_line(ol, units = "metric")

add_mouse_position(ol, projection = "EPSG:4326")

add_overview_map(ol, collapsed = TRUE)
}
\arguments{
\item{ol}{map widget}

\item{units}{units, supported values are
\code{degrees}, \code{imperial}, \code{nautical}, \code{metric} and \code{us}}

\item{projection}{projection used to display coordinates}

\item{collapsed}{start the overview map collapsed?}
}
\description{
Add controls to map
}
\section{Functions}{
\itemize{
\item \code{add_scale_line}: Add scaleline to map

\item \code{add_mouse_position}: Add mouse position to map

\item \code{add_overview_map}: Add overview map to map
}}

\examples{
\dontrun{
  ol() \%>\% add_scale_line()
}

\dontrun{
  ol() \%>\% add_osm_tiles() \%>\% add_mouse_position("EPSG:3857")
}

}
