% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_cartodb_xyz_url}
\alias{get_cartodb_xyz_url}
\title{Get cartodb xyz url.}
\usage{
get_cartodb_xyz_url(layer = "dark_all")
}
\arguments{
\item{layer}{cartodb layer name}
}
\value{
cartodb xyz url for given layer
}
\description{
for available layers please
check \href{https://carto.com/location-data-services/basemaps/}{cartodb-basemaps}
}
