% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tiles.R
\name{guess_vt_format}
\alias{guess_vt_format}
\title{Guess format of vector tiles}
\usage{
guess_vt_format(url)
}
\arguments{
\item{url}{url of the vt service}
}
\description{
Guess format of vector tiles
}
