% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactions}
\alias{interactions}
\alias{add_select}
\alias{add_drag_and_drop}
\title{Add interactions to the map}
\usage{
add_select(ol, condition = "singleClick", display_properties = FALSE)

add_drag_and_drop(ol)
}
\arguments{
\item{ol}{map widget}

\item{condition}{event condition like \code{singleClick}, \code{pointerMove}, ...}

\item{display_properties}{display feature properties on select?}
}
\description{
Add interactions to the map
}
\section{Functions}{
\itemize{
\item \code{add_select}: Add select interaction to the map

\item \code{add_drag_and_drop}: Add drag and drop interaction to the map
}}

