% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{layer_options}
\alias{layer_options}
\title{Layer options}
\usage{
layer_options(opacity = 1, name = NULL, ...)
}
\arguments{
\item{opacity}{opacity of layer}

\item{name}{name of layer}

\item{...}{hidden options}
}
\description{
Layer options
}
