% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{make_icon}
\alias{make_icon}
\title{Make icon}
\usage{
make_icon(filename)
}
\arguments{
\item{filename}{filename of icon}
}
\description{
Encode icon file to base64, so that it can be passed as \code{src} parameter
to \code{\link{marker_style}}.
}
