% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol.R
\name{ol}
\alias{ol}
\title{Create OpenLayers map widget}
\usage{
ol(width = NULL, height = NULL, elementId = NULL,
  options = ol_options())
}
\arguments{
\item{width}{map width}

\item{height}{map height}

\item{elementId}{explicit element id (usually not needed)}

\item{options}{map options, see \code{\link{ol_options}}}
}
\description{
Create OpenLayers map widget
}
