% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ol_options}
\alias{ol_options}
\title{Map options}
\usage{
ol_options(max_zoom_fit = 16, ...)
}
\arguments{
\item{max_zoom_fit}{max zoom when fitting view to extend of data source}

\item{...}{to be added soon}
}
\description{
Map options
}
