% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{openlayers}
\alias{openlayers}
\alias{openlayers-package}
\title{openlayers}
\description{
Makes the open-source JavaScript library \href{https://openlayers.org/}{OpenLayers}
available within R via the \pkg{htmlwidgets} package.
}
\examples{
\dontrun{
  ol() \%>\% add_osm_tiles()

  ol() \%>\% add_stamen_tiles() \%>\%
    add_features(geojsonio::us_cities[1:5, ])
}

}
