% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{select_interaction}
\alias{select_interaction}
\title{Select interaction}
\usage{
select_interaction(condition = "singleClick", property = NULL)
}
\arguments{
\item{condition}{event condition like \code{singleClick}, \code{pointerMove}, ...}

\item{property}{feature property, ...}
}
\description{
Select interaction
}
