% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_view}
\alias{set_view}
\title{Set the view of the map (geographical center and zoom)}
\usage{
set_view(ol, lon = 9.5, lat = 51.31667, zoom = 4)
}
\arguments{
\item{ol}{map widget}

\item{lon}{longitude of center}

\item{lat}{latitude of center}

\item{zoom}{zoom level}
}
\description{
Set the view of the map (geographical center and zoom)
}
