% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R, R/styles_doc.R
\name{stroke_style}
\alias{stroke_style}
\alias{fill_style}
\alias{circle_style}
\alias{marker_style}
\alias{style_methods}
\title{Style methods}
\usage{
stroke_style(color = "green", width = 2)

fill_style(color = "rgba(0, 0, 255, 0.5)")

circle_style(stroke = stroke_style(), fill = fill_style(), radius = 10)

marker_style(src = NULL)
}
\arguments{
\item{color}{stroke or fill color}

\item{width}{stroke width in pixels}

\item{stroke}{stroke style or \code{NULL} (do not stroke)}

\item{fill}{fill style or \code{NULL} (do not fill)}

\item{radius}{radius in pixels}

\item{src}{icon url or \code{NULL} (use default icon)}
}
\description{
style methods corresponding to ol style classes
}
\examples{
  style = stroke_style() + fill_style()

  point_style = circle_style(stroke_style(color = "blue"), fill = NULL)

}
