% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt.R
\name{add_mapbox_vector_tiles}
\alias{add_mapbox_vector_tiles}
\title{Add mapbox vector tiles (mvt) to the map}
\usage{
add_mapbox_vector_tiles(ol, url, attribution = NULL, style = NULL)
}
\arguments{
\item{ol}{map widget}

\item{url}{url of the mvt service}

\item{attribution}{attribution}

\item{style}{style object or \code{NULL} (use default style)}
}
\description{
If you want to create and serve your own vector tiles
you may check \href{http://t-rex.tileserver.ch/}{t-rex}.
}
\examples{
\dontrun{
  url <- "http://tile.mapzen.com/mapzen/vector/v1/water/{z}/{x}/{y}.mvt?api_key=<api_key>"
  attribution <- "&copy; <a href='https://mapzen.com/'>Mapzen</a>"

 ol() \%>\%
   add_mapbox_vector_tiles(url, attribution)
}

}
