% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvt.R, R/vector_tiles.R
\name{add_mapbox_vector_tiles}
\alias{add_mapbox_vector_tiles}
\alias{add_vector_tiles}
\alias{add_topojson_vector_tiles}
\alias{add_geojson_vector_tiles}
\title{Add vector tiles to the map}
\usage{
add_mapbox_vector_tiles(ol, url, attribution = NULL, style = NULL,
  options = layer_options())

add_topojson_vector_tiles(ol, url, attribution = NULL, style = NULL,
  options = layer_options())

add_geojson_vector_tiles(ol, url, attribution = NULL, style = NULL,
  options = layer_options())
}
\arguments{
\item{ol}{map widget}

\item{url}{url of the vector tiles service}

\item{attribution}{attribution}

\item{style}{style object or \code{NULL} (use default style)}

\item{options}{general layer options, see \code{\link{layer_options}}}
}
\description{
If you want to create and serve your own vector tiles
you may check \href{http://t-rex.tileserver.ch/}{t-rex}.
}
\section{Functions}{
\itemize{
\item \code{add_mapbox_vector_tiles}: Add mapbox vector tiles (mvt) to the map

\item \code{add_topojson_vector_tiles}: Add topojson vector tiles to the map

\item \code{add_geojson_vector_tiles}: Add geojson vector tiles to the map
}}

\examples{
\dontrun{
  # mapbox
  key <- "your_access_token"
  url <- paste0(
    "https://{a-d}.tiles.mapbox.com/v4/",
    "mapbox.mapbox-streets-v7/",
    "{z}/{x}/{y}.vector.pbf",
    "?access_token=", key
  )
  attribution <- "Add mapbox attribution here"

  ol() \%>\%
    add_mapbox_vector_tiles(url, attribution)

  # style the tiles
  ol() \%>\%
    add_mapbox_vector_tiles(
      url,
      attribution,
      style = stroke_style(color = "green") + fill_style()
    ) \%>\%
    set_view(zoom = 14)
}

\dontrun{
  # mapzen
  key <- "api_key"
  url <- paste0(
    "https://tile.mapzen.com/mapzen/vector/v1/",
    "roads/{z}/{x}/{y}.topojson",
    "?api_key=", key)
  )

  ol() \%>\%
    add_topojson_vector_tiles(url, "Add mapzen attribution") \%>\%
    set_view(zoom = 16)
}

}
