% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wms.R
\name{add_wms_tiles}
\alias{add_wms_tiles}
\title{Add wms tile layer}
\usage{
add_wms_tiles(ol, url, layers)
}
\arguments{
\item{ol}{map widget}

\item{layers}{layers}
}
\description{
Add wms tile layer
}
\examples{
\dontrun{
  wms_url <- "https://ahocevar.com/geoserver/wms"

  ol() \%>\%
    add_wms_tiles(wms_url, "ne:NE1_HR_LC_SR_W_DR")
}

}
