% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{controls}
\alias{controls}
\alias{add_scale_line}
\title{Add controls to map}
\usage{
add_scale_line(ol, units = "metric")
}
\arguments{
\item{ol}{map widget}

\item{units}{units, supported values are
\code{degrees}, \code{imperial}, \code{nautical}, \code{metric} and \code{us}}
}
\description{
Add controls to map
}
\examples{
\dontrun{
  ol() \%>\% add_scale_line()
}

}
