% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{layer_options}
\alias{layer_options}
\title{Layer options}
\usage{
layer_options(opacity = 1, name = NULL, ...)
}
\arguments{
\item{opacity}{layer opacity}

\item{name}{layer name}

\item{...}{hidden options, e. g. pass \code{type = "base"} to set base layers
used by the layer-switcher control}
}
\description{
Layer options
}
