% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{layer_options}
\alias{layer_options}
\title{Layer options.}
\usage{
layer_options(opacity = 1, select = NULL)
}
\arguments{
\item{opacity}{opacity}

\item{select}{\code{\link{select_interaction}} object}
}
\description{
Layer options.
}
