% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ol_options}
\alias{ol_options}
\title{Map options}
\usage{
ol_options(min_zoom = 0, max_zoom = NULL, max_zoom_fit = 16, ...)
}
\arguments{
\item{min_zoom}{minimal zoom of the map}

\item{max_zoom}{maximal zoom of the map or \code{NULL} (no limit)}

\item{max_zoom_fit}{maximal zoom when fitting the view to the extend of the data source}

\item{...}{hidden options, e. g. pass \code{renderer = "webgl"} to use
webgl as display engine (should not be used when viewing output directly in RStudio)}
}
\description{
Map options
}
