% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R, R/styles_doc.R
\name{stroke_style}
\alias{stroke_style}
\alias{fill_style}
\alias{circle_style}
\alias{marker_style}
\alias{icon_style}
\alias{text_style}
\alias{style_methods}
\title{Create styles}
\usage{
stroke_style(color = "green", width = 2)

fill_style(color = "rgba(0, 0, 255, 0.5)")

circle_style(stroke = stroke_style(), fill = fill_style(), radius = 10)

marker_style(src = NULL, anchor = c(0.5, 0.8), color = NULL)

icon_style(src = NULL, anchor = c(0.5, 0.8), color = NULL)

text_style(text = NULL, property = NULL, scale = 1.5, color = "black",
  stroke = NULL)
}
\arguments{
\item{color}{color to tint the icon or \code{NULL} (keep original)}

\item{width}{stroke width in pixels}

\item{stroke}{stroke style or \code{NULL} (do not stroke)}

\item{fill}{fill style or \code{NULL} (do not fill)}

\item{radius}{radius in pixels}

\item{src}{url of the icon, base64 encoded icon (see \code{\link{make_icon}})
or \code{NULL} (use default icon)}

\item{anchor}{center of the icon relative to its top left corner}

\item{text}{text content, ignored in case \code{property} parameter is set}

\item{property}{feature property used as text content}

\item{scale}{text scale}

\item{color}{stroke or fill color}
}
\value{
style object
}
\description{
Style methods to create style objects corresponding to
\href{http://openlayers.org/en/latest/apidoc/ol.style.html}{openlayers style classes}.
}
\note{
\code{color} (in \code{fill_style}), \code{radius} and \code{text} can be single values applied to all features
  or vectors of length equal to the number of features
}
\examples{
  style = stroke_style() + fill_style()

  point_style = circle_style(stroke_style(color = "blue"), fill = NULL)
}
